Fixed cucumber incompatibility and some features
Cucumber  defines a  `#before` method  in main  that accepts  an argument  and a
block. Sinatra  also defines a  `#before` method in  main that accepts  a block.
When `before {}`  is called in the  sinatra file, cucumber tries  it against its
own     `#before`     method     rather    than     sinatra's,     hence     the
`Sinatra::Application.before` is needed